backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=9.5.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
gdown>=5.1.0
anthropic>=0.28.0
